+++ types/broccoli-plugin/index.d.ts
@@ -1,0 +1,11 @@
+declare module 'broccoli-plugin' {
+
+  export interface Tree {}
+
+  export default class Plugin {
+    constructor(inputTrees: [Tree], options: any)
+    inputPaths: [string];
+    outputPath: string
+  }
+
+}